# dollmerchant/docs/swagger/swagger.yaml
basePath: /dev/mch/v1
host: 192.168.0.167:9555
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 娃娃机商户平台 Api文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 娃娃机商户平台 Api文档
  version: "1.0"
paths:
...
  /notice/add:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 公告名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增公告
      tags:
      - Notice
  /notice/del:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 公告ID
        in: formData
        name: Notice_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除公告
      tags:
      - Notice
  /notice/info:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 公告ID
        in: query
        name: Notice_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取公告
      tags:
      - Notice
  /notice/list:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 偏移量
        in: query
        name: offset
        required: true
        type: string
      - description: 总条数
        in: query
        name: total_size
        required: true
        type: string
      - description: 一页多少条数据
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取公告列表
      tags:
      - Notice
  /notice/upd:
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 公告ID
        in: formData
        name: Notice_id
        required: true
        type: string
      - description: 公告名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 更新公告
      tags:
      - Notice
  /place/add:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 投放地点名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增投放地点
      tags:
      - Place
  /place/del:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 投放地点ID
        in: formData
        name: place_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除投放地点
      tags:
      - Place
  /place/info:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 投放地点ID
        in: query
        name: Place_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取投放地点
      tags:
      - Place
  /place/list:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 偏移量
        in: query
        name: offset
        required: true
        type: string
      - description: 总条数
        in: query
        name: total_size
        required: true
        type: string
      - description: 一页多少条数据
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取投放地点列表
      tags:
      - Place
  /place/upd:
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 投放地点ID
        in: formData
        name: Place_id
        required: true
        type: string
      - description: 投放地点名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 更新投放地点
      tags:
      - Place
...
swagger: "2.0"